backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cohere>=5.20.1",
    "fastapi>=0.115.6",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "uvicorn>=0.36.2",
]
